refactor(awto-es): inline `macros` export
