+++ awto-es/src/lib.rs
@@ -1,17 +1,21 @@
-use awto::{Awto, AwtoBuilder};
-pub use awto_es_macros as macros;
+use awto::*;
+#[doc(hidden)]
+pub use awto_es_macros::*;
+#[doc(inline)]
 pub use command::*;
 pub use error::Error;
-pub use message::*;
+pub use topic::*;
+#[doc(inline)]
 pub use query::*;
 
-pub mod awto;
-pub mod command;
+mod awto;
+mod command;
 mod error;
-mod message;
+mod topic;
 pub mod postgres;
-pub mod query;
+mod query;
 
+/// Builds an awto app.
 pub fn build<ES>(event_store: ES, redpanda_host: impl Into<String> + Clone) -> AwtoBuilder<ES>
 where
     ES: EventStore + Clone + Send + Sync + Unpin + 'static,
